add CDP comunicator
